+++ tests/pytests/test_lro_drivers.py
@@ -115,4 +115,3 @@ def test_lro_creation(lro_lroclabel):
     with LrocPds3Driver(lro_lroclabel) as m:
         d = m.to_dict()
         assert isinstance(d, dict)
-        assert(set(d.keys()) == m.required_keys)
