+++ tests/pytests/test_mro_drivers.py
@@ -66,4 +66,3 @@ def test_ctx_creation(mroctx_label):
     with CtxPds3Driver(mroctx_label) as m:
         d = m.to_dict()
     assert isinstance(d, dict)
-    assert(set(d.keys()) == m.required_keys)
